backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
grpcio==1.67.0
grpcio-tools==1.67.0
Pillow==10.0.0
protobuf==5.28.2